requirements.txt
python-telegram-bot[socks] == 11.1.0 # Python API for Telegram Bot
